requirements.txt
asgiref
backports.zoneinfo
dj-database-url
django
djangorestframework
gunicorn
pytz
sqlparse
whitenoise
numpy